Fixed table responsiveness and navigation roles of pharmacist and cashier

+++ medicine/medicine_list_page.php
@@ -1,5 +1,7 @@
 <?php
-session_start();
+if (session_status() == PHP_SESSION_NONE) {
+  session_start();
+}
 include "./controllers/retrieve-list.php";
 include "../assets/components/side-bar.php";
 include "../assets/components/banner.php";
@@ -32,17 +34,19 @@ include "../assets/components/banner.php";
         </svg>
       </button>
     </form>
-
-    <div class="header-right">
-      <a href="/pharma-suite/medicine/medicine_add_page.php" class="add-btn">
-        <svg class="add-icon" viewBox="0 0 24 24">
-          <path d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z" />
-        </svg>
-      </a>
-    </div>
+    <?php
+    if (strtoupper($_SESSION['position']) != "CASHIER")
+      echo '<div class="header-right">
+              <a href="/pharma-suite/medicine/medicine_add_page.php" class="add-btn">
+                <svg class="add-icon" viewBox="0 0 24 24">
+                  <path d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z" />
+                </svg>
+              </a>
+          </div>';
+    ?>
   </header>
   <main class="list-main">
-    <table cellspacing="3">
+    <table cellspacing="3" class="list-table">
       <tr>
         <td class="list-row1">Name</td>
         <td class="list-row1">Catagory</td>
@@ -50,7 +54,10 @@ include "../assets/components/banner.php";
         <td class="list-row1">Quantity</td>
         <td class="list-row1">Unit Price</td>
         <td class="list-row1">Expiration date</td>
-        <td class="list-row1">Actions</td>
+        <?php
+        if (strtoupper($_SESSION['position']) != "CASHIER")
+          echo '<td class="list-row1">Actions</td>'
+        ?>
       </tr>
       <?php
       if (empty($medicines))
@@ -66,11 +73,14 @@ include "../assets/components/banner.php";
                   <td class='list-cells'>" . $row['unit_price'] . "</td>
                   <td class='list-cells " . ($row['expiry_date'] < date('Y-m-d') ? 'expired' : '') . "'>
                       " . $row['expiry_date'] . "
-                  </td>
+                  </td>";
+          if (strtoupper($_SESSION['position']) != "CASHIER")
+            echo "
                   <td>
                     <a style='background-color: #55cc55' class='action-btn' href='./medicine_edit_page.php?id=" . $row['id'] . "'>Edit</a>
                     <a style='background-color: #cc5555' class='action-btn' href='./controllers/delete-item.php?id=" . $row['id'] . "' onclick='return confirmDelete(" . $row['stock_quantity'] . ")'>Delete</a>
-                  </td>
+                  </td>";
+          echo "
               </tr>
           ";
         }
