+++ transaction/transaction_list_page.php
@@ -1,5 +1,7 @@
 <?php
-session_start();
+if (session_status() == PHP_SESSION_NONE) {
+  session_start();
+}
 include "./controllers/retrieve-list.php";
 include "../assets/components/side-bar.php";
 include "../assets/components/banner.php";
@@ -32,14 +34,17 @@ include "../assets/components/banner.php";
         </svg>
       </button>
     </form>
-
-    <div class="header-right">
-      <a href="/pharma-suite/transaction/transaction_add_page.php" class="add-btn">
-        <svg class="add-icon" viewBox="0 0 24 24">
-          <path d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z" />
-        </svg>
-      </a>
-    </div>
+    <?php
+    if (strtoupper($_SESSION['position']) != "PHARMACIST")
+      echo '
+        <div class="header-right">
+          <a href="/pharma-suite/transaction/transaction_add_page.php" class="add-btn">
+            <svg class="add-icon" viewBox="0 0 24 24">
+              <path d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z" />
+            </svg>
+          </a>
+        </div>';
+    ?>
   </header>
   <main class="list-main">
     <table cellspacing="3" class="list-table">
@@ -51,7 +56,10 @@ include "../assets/components/banner.php";
         <td class="list-row1">Quantity</td>
         <td class="list-row1">Total Price</td>
         <td class="list-row1">Date</td>
-        <td class="list-row1">Actions</td>
+        <?php
+        if (strtoupper($_SESSION['position']) != "PHARMACIST")
+          echo '<td class="list-row1">Actions</td>';
+        ?>
       </tr>
       <?php
       if (empty($transactions))
@@ -67,10 +75,15 @@ include "../assets/components/banner.php";
                   <td class='list-cells'>" . $row['quantity'] . "</td>
                   <td class='list-cells'>" . ($row['quantity'] * $row['unit_price']) . " birr</td>
                   <td class='list-cells'>" . date('Y-m-d', strtotime($row['transaction_date'])) . "</td>
+                  ";
+          if (strtoupper($_SESSION['position']) != "PHARMACIST")
+            echo "
                   <td>
                     <a style='background-color: #55cc55' class='action-btn' href='./transaction_edit_page.php?id=" . $row['id'] . "'>Edit</a>
                     <a style='background-color: #cc5555' class='action-btn' href='./controllers/delete-item.php?id=" . $row['id'] . "' onclick='return confirmDelete(" . $row['stock_quantity'] . ")'>Delete</a>
                   </td>
+                  ";
+          echo "
               </tr>
           ";
         }
